+++ SpringApp/src/main/java/Main.java
@@ -10,8 +10,8 @@ public class Main {
     private static final AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(Config.class);
 
     public static void main(String[] args) {
-        List<Question> questions = context.getBean("csv", CSV.class).getQuestionsRel("Example1.csv");
-        for (Question question: questions) System.out.println(question.toString());
+        List<Question> questions = ((CSV) context.getBean("csv")).getQuestions("Example1.csv");
+        System.out.println(questions);
         context.close();
     }
 }
